feat: add reusable device selector
