+++ packages/ui/src/components/cms/DeviceSelector.tsx
@@ -1,19 +1,13 @@
 import { ReloadIcon } from "@radix-ui/react-icons";
-import {
-  Button,
-  Select,
-  SelectTrigger,
-  SelectValue,
-  SelectContent,
-  SelectItem,
-} from "../atoms/shadcn";
-import { devicePresets } from "@ui/utils/devicePresets";
+import { Button } from "../atoms/shadcn";
+import DeviceSelectorCommon from "../common/DeviceSelector";
 
 interface Props {
   deviceId: string;
   orientation: "portrait" | "landscape";
   setDeviceId: (id: string) => void;
   toggleOrientation: () => void;
+  showLegacyButtons?: boolean;
 }
 
 export default function DeviceSelector({
@@ -21,21 +15,15 @@ export default function DeviceSelector({
   orientation,
   setDeviceId,
   toggleOrientation,
+  showLegacyButtons,
 }: Props): React.JSX.Element {
   return (
     <div className="flex items-center gap-2">
-      <Select value={deviceId} onValueChange={setDeviceId}>
-        <SelectTrigger aria-label="Device" className="w-40">
-          <SelectValue />
-        </SelectTrigger>
-        <SelectContent>
-          {devicePresets.map((p) => (
-            <SelectItem key={p.id} value={p.id}>
-              {p.label}
-            </SelectItem>
-          ))}
-        </SelectContent>
-      </Select>
+      <DeviceSelectorCommon
+        deviceId={deviceId}
+        onChange={setDeviceId}
+        showLegacyButtons={showLegacyButtons}
+      />
       <Button
         variant="outline"
         onClick={toggleOrientation}
@@ -48,3 +36,4 @@ export default function DeviceSelector({
     </div>
   );
 }
+
