+++ packages/ui/src/components/cms/page-builder/PageToolbar.tsx
@@ -1,11 +1,9 @@
 import type { Locale } from "@/i18n/locales";
-import { DesktopIcon, LaptopIcon, MobileIcon } from "@radix-ui/react-icons";
+import { ReloadIcon } from "@radix-ui/react-icons";
 import { Button, Input } from "../../atoms/shadcn";
-import { getLegacyPreset } from "@ui/utils/devicePresets";
-import DeviceSelector from "../DeviceSelector";
+import DeviceSelector from "../../common/DeviceSelector";
 
 interface Props {
-  viewport: "desktop" | "tablet" | "mobile";
   deviceId: string;
   setDeviceId: (id: string) => void;
   orientation: "portrait" | "landscape";
@@ -22,7 +20,6 @@ interface Props {
 }
 
 const PageToolbar = ({
-  viewport,
   deviceId,
   setDeviceId,
   locale,
@@ -34,38 +31,28 @@ const PageToolbar = ({
   toggleGrid,
   gridCols,
   setGridCols,
+  orientation,
+  setOrientation,
 }: Props) => (
   <div className="flex flex-col gap-4">
     <div className="flex justify-end gap-2">
-      {(["desktop", "tablet", "mobile"] as const).map((t) => {
-        const preset = getLegacyPreset(t);
-        const Icon =
-          t === "desktop" ? DesktopIcon : t === "tablet" ? LaptopIcon : MobileIcon;
-        return (
-          <Button
-            key={t}
-            variant={deviceId === preset.id ? "default" : "outline"}
-            onClick={() => setDeviceId(preset.id)}
-            aria-label={t}
-          >
-            <Icon />
-            <span className="sr-only">
-              {t.charAt(0).toUpperCase() + t.slice(1)}
-            </span>
-          </Button>
-        );
-      })}
       <DeviceSelector
         deviceId={deviceId}
-        orientation={orientation}
-        setDeviceId={(id) => {
+        onChange={(id) => {
           setDeviceId(id);
           setOrientation("portrait");
         }}
-        toggleOrientation={() =>
+        showLegacyButtons
+      />
+      <Button
+        variant="outline"
+        onClick={() =>
           setOrientation((o) => (o === "portrait" ? "landscape" : "portrait"))
         }
-      />
+        aria-label="Rotate"
+      >
+        <ReloadIcon className={orientation === "landscape" ? "rotate-90" : ""} />
+      </Button>
     </div>
     <div className="flex items-center justify-end gap-2">
       <Button
